TST: only remove old link if it's there

+++ test-setup.sh
@@ -1,7 +1,9 @@
 python setup.py build
 LIB=`find build -name pyca.so`
 LINK='pyca.so'
-rm $LINK
+if [ -L $LINK ]; then
+  rm $LINK
+fi
 if [ -z $LIB ]; then
   exit
 fi
